practice-app/backend/fithub/requirements.txt
Django==5.2
PyMySQL==1.1.1                           # For MySQL connection if mysqlclient won't work
python-dotenv==1.0.1                     # To read .env file
drf-yasg>=1.21.7                         # For swagger documentation
djangorestframework_simplejwt==5.5.0     # For JWT authentication
django-cors-headers==4.7.0               # For CORS headerspip install
cryptography==44.0.3
requests==2.28.1                         #for wikidata connection
jsonschema==4.17.0

# Cloudinary integration (recipe optional image upload to cloud)
cloudinary>=1.41
django-cloudinary-storage>=0.3.0
Pillow>=10.0  # needed for ImageField
django-extensions
Werkzeug
pyOpenSSL
django-extensions